requirements.txt
requests==2.31.0
pandas==1.5.3
python-dotenv==1.0.0  # Only if you're using .env files
